buffer::buffer

Struct BufferManager

pub struct BufferManager<F: FileManager> {
    file_manager: Arc<F>,
    page_map: RwLock<HashMap<PageId, Arc<PageEntry>>>,
    frames: Vec<BufferFrame>,
}

The buffer manager responsible for handling the cache pool of data pages.

Fields
file_manager: Arc<F>page_map: RwLock<HashMap<PageId, Arc<PageEntry>>>frames: Vec<BufferFrame>
Implementations

impl<F: FileManager> BufferManager<F>

pub fn new(file_manager: Arc<F>, pool_size: usize) -> Self

Creates a new empty buffer manager. Allocates a predefined number of buffer frames. Sets up internal structures required for managing the pool.

pub fn read_page( &self, page_id: PageId, ) -> Result<PageReadGuard<'_>, BufferError>

Retrieves a page from the buffer pool based on its page ID. If the page cannot be found in the buffer, it is first loaded from disk, cached, then returned.

Params

page_id: The identifier of the page to be retrieved.

Returns

A Result where the Ok contains a PageReadGuard. A PageReadGuard encapsulates the latch needed to access the underlying &Page.

fn claim_free_frame(&self, for_page_id: PageId) -> Option<usize>

Goes through the frames to find an empty one that can be used This is done by iterating over the vector, probing for a write-latch without waiting if it is already taken. This ensures already-in-use frames are skipped. When a free frame is found (frame.page_id = None), the page_id is set to the provided one This ensures the frame is claimed as taken, even if the page has not yet been loaded

Params

page_id: the ID of the page we intend to load into the free frame

Returns

The FrameID for the identified and claimed free frame. If no empty frame is found, None.

fn read_guard_from_frame(&self, frame_id: usize) -> PageReadGuard<'_>

Computes a PageReadGuard object for a frame.

fn wait_until_ready(entry: &Arc<PageEntry>) -> usize

Waits for the Mutex on a PageEntry to be free to access and the page is loaded into memory (PageState = Ready(FrameId))